Add files via upload
Wrote a class Tweet and the main-class for the conversion. In both there are parts commented out which later have to be included again.
TwitterConvert currently does not work, because the (json-)parsing of the string fails and returns null. When this is fixed, the other
properties of the tweets have to be read to. Furthermore the tweets have to be filtered with .filter(t->t.getHashtags().size() > 0).